various: split settings and home sections
